CIFAR10/requirements.txt
matplotlib==3.5.2
numpy==1.22.3
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.6.0.66
Pillow==9.5.0
torch==1.12.0
torchsummary==1.5.1
torchvision==0.13.0
